WIP start moving AnalysisStorage
